Address review: de-dup GC barrier, harden commit insert, tune GC allocations
- M1: move compute_cleanup_ts/run_gc_dirty_inner/run_gc_full onto Inner so
  the background gc worker and the manual run_gc API share one copy of the
  SeqCst publish-fence-rescan barrier instead of duplicating it verbatim.
- M2: commit apply uses get_or_insert_with + an is_removed re-check under
  the write lock instead of SkipMap::insert (which replaces), so a
  concurrent writer's node can't be clobbered; the seeded chain avoids a
  reclaimable empty node mid-insert.
- m1: Versions::drain only shrinks when capacity outgrows 2x the live set,
  avoiding allocation thrash on hot keys under frequent gc.
- m2: run_gc_dirty_inner de-duplicates keys within a drain pass.

Build + clippy clean; full suite (395 tests) green.
